social_media_analytics/requirements.txt
textblob
vaderSentiment
wordcloud
matplotlib
pygal
Pillow
opencv-python
SpeechRecognition
pydub
emoji
pandas
numpy
nltk
seaborn
flask
plotly
transformers
torch
requests
